feat: improve chat performance
